userService/requirements.txt
alembic==1.13.2
bcrypt==4.1.3
blinker==1.8.2
cffi==1.16.0
click==8.1.7
cryptography==43.0.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.1.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
Werkzeug==3.0.3
